virt_mshv_vtl: Refactor deliver_synic_messages
